requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Email processing
secure-smtplib>=0.1.1

# Azure dependencies
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-search-documents>=11.4.0

# Development dependencies
pytest>=7.4.0
black>=23.12.0
flake8>=6.1.0